[git-p4: depot-paths = "//open/dy/dev/": change = 14356]

+++ luciddb/initBuild.sh
@@ -164,7 +164,9 @@ if $without_dist_build ; then
     echo Skipping distribution build.
 else
     cd ${luciddb_dir}/../farrago
-    cp -f ./dist/ExampleRelease.properties ./dist/FarragoRelease.properties
+    if [ ! -e ./dist/FarragoRelease.properties ]; then
+        cp -f ./dist/ExampleRelease.properties ./dist/FarragoRelease.properties
+    fi
     cd ${luciddb_dir}
     ./distBuild.sh --without-init-build ${FARRAGO_DIST_FLAGS}
 fi
